refactor sql utils class to build upsert transaction
